+++ example/lib/webservice/login/login_webservice.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
-import 'package:fgen/webservice/login/login_service.dart';
 import 'package:injectable/injectable.dart';
 import 'package:retrofit/retrofit.dart';
+import 'package:example/webservice/login/login_service.dart';
 
 part 'login_webservice.g.dart';
 
@@ -10,4 +10,7 @@ part 'login_webservice.g.dart';
 abstract class LoginWebService extends LoginService {
   @factoryMethod
   factory LoginWebService(Dio dio) = _LoginWebService;
+
+  @GET('/example')
+  Future<void> getExample();
 }