+++ lib/fgen.dart
@@ -1,24 +1,43 @@
 import 'dart:io';
 
+import 'package:fgen/src/code/fgen_template.dart';
 import 'package:fgen/src/code/generate_code.dart';
 import 'package:fgen/src/util/pubspec_utils.dart';
 
 /// Generate code for a given name, returns 0 if successful, -1 if not
-int generateCode({
+Future<int> generateCode({
   required String name,
-  required Set<GenerateType> types,
-}) {
+  required List<FgenTemplate> templatesToUse,
+}) async {
   if (name.isEmpty) {
     print('Name cannot be empty');
     return -1;
   }
-  if (types.isEmpty) {
+  if (templatesToUse.isEmpty) {
     print('No code to generate');
     return -1;
   }
   final pubspec = PubspecUtils.getPubspecFromFolder(Directory.current.path);
-  for (final type in types) {
-    GenerateCode.write(type, name, pubspec, types);
+  var exitCode = 0;
+  for (final template in templatesToUse) {
+    if (template.templateFiles.isEmpty) {
+      print('ERROR: No template files found for ${template.name}');
+      exitCode = 2;
+      continue;
+    }
+    for (final templateFile in template.templateFiles.entries) {
+      final success = await CodeGenerator.generate(
+        name: name,
+        template: template,
+        templateFile: templateFile.key,
+        fileSuffix: templateFile.value,
+        pubspec: pubspec,
+        usedTemplates: templatesToUse.map((e) => e.name).toList(),
+      );
+      if (success != 0) {
+        exitCode = success;
+      }
+    }
   }
-  return 0;
+  return exitCode;
 }