+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## 1.0.0 - 2024/01/02
+
+- Use `flutter_navigation_generator` instead of `get_x_navigation_generator`
+- Changed to template based generation
+- Added support for custom templates
+- Added support for custom options
+
 ## 0.0.3 - 2023/04/14
 
 - Replaced GetIt.I with getIt
-- Fix package name for get_x_navigation_generator_annotations
+- Fix package name for flutter_navigation_generator_annotations
 
 ## 0.0.2 - 2023/01/07